requirements.txt
fastapi
uvicorn
python-docx
pymupdf
sentence-transformers
faiss-cpu
python-multipart
